+++ Release/CHANGELOG.md
@@ -1,3 +1,18 @@
+# 1.0.5
+
+Added config options
+
+Added Risk of Options support
+
+Added Risky Mod support
+
+No longer requires Dronemeld to have the Lemurian master name in the config
+
+Added quality of life changes (all configurable)
+
+Added some experimental gameplay changes for how the inventories are managed (Experimental tab in config)
+
+
 # 1.0.4
 
 fixed lemur inventory not updating on summon
@@ -5,7 +20,9 @@ fixed lemur inventory not updating on summon
 # 1.0.3
 
 supports modded elites
+
 lemurians will drop 1 scrap for each unique item that they had on death.
+
 lemurians will now switch between the various high tier elite types once they are fully evolved
 
 # 1.0.2